gcsuploader/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.11.1
firebase-admin==6.2.0